fix: refactor termwrap init lifecycle to eliminate terminal race condition
Opening a terminal pane sometimes showed a permanent black screen because
the backend PTY was spawned (via resyncController) before the frontend
subscribed to the data stream. Shell prompt output arriving in this window
was either dropped (no subscriber) or buffered in heldData and never
flushed to xterm.js.

Restructure TermWrap into a strict 3-phase lifecycle:
1. CONSTRUCT (sync) - create Terminal, load addons, register OSC handlers
2. INIT (async) - subscribe → load data → flush heldData → resync controller
3. RUNNING - resize, receive data, periodic cache

Key changes:
- Subscribe to file subject BEFORE calling resyncController (which spawns PTY)
- Flush heldData after loadInitialTerminalData completes
- Remove resyncController side-effect from handleResize
- Extract OSC handlers to termosc.ts (~220 lines)
- Extract agent registration to termagent.ts (~80 lines)
- termwrap.ts reduced from 781 to ~430 lines
