+++ frontend/app/view/term/termwrap.ts
@@ -7,7 +7,6 @@ import { RpcApi } from "@/app/store/wshclientapi";
 import { TabRpcClient } from "@/app/store/wshrpcutil";
 import { WOS, atoms, fetchWaveFile, getSettingsKeyAtom, globalStore, openLink } from "@/app/store/global";
 import * as services from "@/app/store/services";
-import { getWebServerEndpoint } from "@/util/endpoints";
 import { PLATFORM, PlatformMacOS } from "@/util/platformutil";
 import { base64ToArray, fireAndForget } from "@/util/util";
 import { SearchAddon } from "@xterm/addon-search";
@@ -22,6 +21,8 @@ import debug from "debug";
 import { debounce } from "throttle-debounce";
 import { FilePathLinkProvider, makeFilePathHandler } from "./filelinkprovider";
 import { FitAddon } from "./fitaddon";
+import { registeredAgentsByBlock, unregisterAgent } from "./termagent";
+import { handleOsc7Command, handleOsc16162Command, handleOscTitleCommand, handleOscWaveCommand } from "./termosc";
 
 const dlog = debug("wave:termwrap");
 
@@ -29,7 +30,6 @@ const TermFileName = "term";
 const TermCacheFileName = "cache:term:full";
 const MinDataProcessedForCache = 100 * 1024;
 
-// detect webgl support
 function detectWebGLSupport(): boolean {
     try {
         const canvas = document.createElement("canvas");
@@ -49,399 +49,19 @@ type TermWrapOptions = {
     sendDataHandler?: (data: string) => void;
 };
 
-// Track registered agent IDs per block to detect changes
-const registeredAgentsByBlock = new Map<string, string>();
-
-// Register an agent with the reactive messaging backend
-async function registerAgent(agentId: string, blockId: string, tabId?: string): Promise<void> {
-    try {
-        const url = getWebServerEndpoint() + "/wave/reactive/register";
-        const response = await fetch(url, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-                agent_id: agentId,
-                block_id: blockId,
-                tab_id: tabId || "",
-            }),
-        });
-        if (!response.ok) {
-            let errorMsg = `HTTP ${response.status}`;
-            try {
-                const data = await response.json();
-                errorMsg = data.error || errorMsg;
-            } catch {
-                // Response body not JSON, use status
-            }
-            console.error("[reactive] failed to register agent:", errorMsg);
-        } else {
-            console.log("[reactive] registered agent", agentId, "->", blockId);
-        }
-    } catch (e) {
-        console.error("[reactive] error registering agent:", e);
-    }
-}
-
-// Unregister an agent from the reactive messaging backend
-async function unregisterAgent(agentId: string): Promise<void> {
-    try {
-        const url = getWebServerEndpoint() + "/wave/reactive/unregister";
-        const response = await fetch(url, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ agent_id: agentId }),
-        });
-        if (!response.ok) {
-            let errorMsg = `HTTP ${response.status}`;
-            try {
-                const data = await response.json();
-                errorMsg = data.error || errorMsg;
-            } catch {
-                // Response body not JSON, use status
-            }
-            console.error("[reactive] failed to unregister agent:", errorMsg);
-        } else {
-            console.log("[reactive] unregistered agent", agentId);
-        }
-    } catch (e) {
-        console.error("[reactive] error unregistering agent:", e);
-    }
-}
-
-// Handle agent ID changes from environment variables
-function handleAgentIdChange(blockId: string, newAgentId: string | undefined, tabId?: string): void {
-    const previousAgentId = registeredAgentsByBlock.get(blockId);
-
-    // No change
-    if (previousAgentId === newAgentId) {
-        return;
-    }
-
-    // Unregister previous agent if there was one
-    if (previousAgentId) {
-        fireAndForget(() => unregisterAgent(previousAgentId));
-        registeredAgentsByBlock.delete(blockId);
-    }
-
-    // Register new agent if there is one
-    if (newAgentId) {
-        registeredAgentsByBlock.set(blockId, newAgentId);
-        fireAndForget(() => registerAgent(newAgentId, blockId, tabId));
-    }
-}
-
-function handleOscWaveCommand(data: string, blockId: string, loaded: boolean): boolean {
-    if (!loaded) {
-        return true;
-    }
-    if (!data || data.length === 0) {
-        console.log("Invalid Wave OSC command received (empty)");
-        return true;
-    }
-
-    // Expected formats:
-    // "setmeta;{JSONDATA}"
-    // "setmeta;[wave-id];{JSONDATA}"
-    const parts = data.split(";");
-    if (parts[0] !== "setmeta") {
-        console.log("Invalid Wave OSC command received (bad command)", data);
-        return true;
-    }
-    let jsonPayload: string;
-    let waveId: string | undefined;
-    if (parts.length === 2) {
-        jsonPayload = parts[1];
-    } else if (parts.length >= 3) {
-        waveId = parts[1];
-        jsonPayload = parts.slice(2).join(";");
-    } else {
-        console.log("Invalid Wave OSC command received (1 part)", data);
-        return true;
-    }
-
-    let meta: any;
-    try {
-        meta = JSON.parse(jsonPayload);
-    } catch (e) {
-        console.error("Invalid JSON in Wave OSC command:", e);
-        return true;
-    }
-
-    if (waveId) {
-        // Resolve the wave id to an ORef using our ResolveIdsCommand.
-        fireAndForget(() => {
-            return RpcApi.ResolveIdsCommand(TabRpcClient, { blockid: blockId, ids: [waveId] })
-                .then((response: { resolvedids: { [key: string]: any } }) => {
-                    const oref = response.resolvedids[waveId];
-                    if (!oref) {
-                        console.error("Failed to resolve wave id:", waveId);
-                        return;
-                    }
-                    services.ObjectService.UpdateObjectMeta(oref, meta);
-                })
-                .catch((err: any) => {
-                    console.error("Error resolving wave id", waveId, err);
-                });
-        });
-    } else {
-        // No wave id provided; update using the current block id.
-        fireAndForget(() => {
-            return services.ObjectService.UpdateObjectMeta(WOS.makeORef("block", blockId), meta);
-        });
-    }
-    return true;
-}
-
-// for xterm handlers, we return true always because we "own" OSC 7.
-// even if it is invalid we dont want to propagate to other handlers
-function handleOsc7Command(data: string, blockId: string, loaded: boolean): boolean {
-    if (!loaded) {
-        return true;
-    }
-    if (data == null || data.length == 0) {
-        console.log("Invalid OSC 7 command received (empty)");
-        return true;
-    }
-    if (data.length > 1024) {
-        console.log("Invalid OSC 7, data length too long", data.length);
-        return true;
-    }
-
-    let pathPart: string;
-    try {
-        const url = new URL(data);
-        if (url.protocol !== "file:") {
-            console.log("Invalid OSC 7 command received (non-file protocol)", data);
-            return true;
-        }
-        pathPart = decodeURIComponent(url.pathname);
-
-        // Normalize double slashes at the beginning to single slash
-        if (pathPart.startsWith("//")) {
-            pathPart = pathPart.substring(1);
-        }
-
-        // Handle Windows paths (e.g., /C:/... or /D:\...)
-        if (/^\/[a-zA-Z]:[\\/]/.test(pathPart)) {
-            // Strip leading slash and normalize to forward slashes
-            pathPart = pathPart.substring(1).replace(/\\/g, "/");
-        }
-    } catch (e) {
-        console.log("Invalid OSC 7 command received (parse error)", data, e);
-        return true;
-    }
-
-    setTimeout(() => {
-        fireAndForget(async () => {
-            await services.ObjectService.UpdateObjectMeta(WOS.makeORef("block", blockId), {
-                "cmd:cwd": pathPart,
-            });
-
-            const rtInfo = { "cmd:hascurcwd": true };
-            const rtInfoData: CommandSetRTInfoData = {
-                oref: WOS.makeORef("block", blockId),
-                data: rtInfo,
-            };
-            await RpcApi.SetRTInfoCommand(TabRpcClient, rtInfoData).catch((e) =>
-                console.log("error setting RT info", e)
-            );
-        });
-    }, 0);
-    return true;
-}
-
-// OSC 0/2 - Window Title (used by Claude Code for activity summaries)
-// Debounce map to prevent rapid title updates
-const titleUpdateDebounceMap = new Map<string, ReturnType<typeof setTimeout>>();
-const TITLE_UPDATE_DEBOUNCE_MS = 300;
-
-function handleOscTitleCommand(data: string, blockId: string, loaded: boolean): boolean {
-    if (!loaded) {
-        return false; // Let xterm handle it too for window title
-    }
-    if (data == null || data.length === 0) {
-        return false;
-    }
-    if (data.length > 256) {
-        // Truncate very long titles
-        data = data.substring(0, 256);
-    }
-
-    // Parse the title - strip common prefixes
-    let activity = data;
-    if (activity.startsWith("Claude: ")) {
-        activity = activity.substring(8);
-    } else if (activity.startsWith("Claude Code: ")) {
-        activity = activity.substring(13);
-    }
-
-    // Debounce rapid title updates per block
-    const existingTimeout = titleUpdateDebounceMap.get(blockId);
-    if (existingTimeout) {
-        clearTimeout(existingTimeout);
-    }
-
-    const timeout = setTimeout(() => {
-        titleUpdateDebounceMap.delete(blockId);
-        fireAndForget(async () => {
-            await services.ObjectService.UpdateObjectMeta(WOS.makeORef("block", blockId), {
-                "term:activity": activity,
-            });
-        });
-    }, TITLE_UPDATE_DEBOUNCE_MS);
-
-    titleUpdateDebounceMap.set(blockId, timeout);
-
-    return false; // Return false to let xterm also handle it (for native window title)
-}
-
-// OSC 16162 - Shell Integration Commands
-// See aiprompts/wave-osc-16162.md for full documentation
-type Osc16162Command =
-    | { command: "A"; data: {} }
-    | { command: "C"; data: { cmd64?: string } }
-    | { command: "M"; data: { shell?: string; shellversion?: string; uname?: string } }
-    | { command: "D"; data: { exitcode?: number } }
-    | { command: "I"; data: { inputempty?: boolean } }
-    | { command: "R"; data: {} }
-    | { command: "E"; data: { [key: string]: string } } // Environment variables update
-    | { command: "X"; data: { agentmux_url?: string; agentmux_token?: string } }; // AgentMux config
-
-function handleOsc16162Command(data: string, blockId: string, loaded: boolean, terminal: Terminal): boolean {
-    if (!loaded) {
-        return true;
-    }
-    if (!data || data.length === 0) {
-        return true;
-    }
-
-    const parts = data.split(";");
-    const commandStr = parts[0];
-    const jsonDataStr = parts.length > 1 ? parts.slice(1).join(";") : null;
-    let parsedData: Record<string, any> = {};
-    if (jsonDataStr) {
-        try {
-            parsedData = JSON.parse(jsonDataStr);
-        } catch (e) {
-            console.error("Error parsing OSC 16162 JSON data:", e);
-        }
-    }
-
-    const cmd: Osc16162Command = { command: commandStr, data: parsedData } as Osc16162Command;
-    const rtInfo: ObjRTInfo = {};
-    switch (cmd.command) {
-        case "A":
-            rtInfo["shell:state"] = "ready";
-            break;
-        case "C":
-            rtInfo["shell:state"] = "running-command";
-            if (cmd.data.cmd64) {
-                const decodedLen = Math.ceil(cmd.data.cmd64.length * 0.75);
-                if (decodedLen > 8192) {
-                    rtInfo["shell:lastcmd"] = `# command too large (${decodedLen} bytes)`;
-                } else {
-                    try {
-                        const decodedCmd = atob(cmd.data.cmd64);
-                        rtInfo["shell:lastcmd"] = decodedCmd;
-                    } catch (e) {
-                        console.error("Error decoding cmd64:", e);
-                        rtInfo["shell:lastcmd"] = null;
-                    }
-                }
-            } else {
-                rtInfo["shell:lastcmd"] = null;
-            }
-            break;
-        case "M":
-            if (cmd.data.shell) {
-                rtInfo["shell:type"] = cmd.data.shell;
-            }
-            if (cmd.data.shellversion) {
-                rtInfo["shell:version"] = cmd.data.shellversion;
-            }
-            if (cmd.data.uname) {
-                rtInfo["shell:uname"] = cmd.data.uname;
-            }
-            break;
-        case "D":
-            if (cmd.data.exitcode != null) {
-                rtInfo["shell:lastcmdexitcode"] = cmd.data.exitcode;
-            } else {
-                rtInfo["shell:lastcmdexitcode"] = null;
-            }
-            break;
-        case "I":
-            if (cmd.data.inputempty != null) {
-                rtInfo["shell:inputempty"] = cmd.data.inputempty;
-            }
-            break;
-        case "R":
-            if (terminal.buffer.active.type === "alternate") {
-                terminal.write("\x1b[?1049l");
-            }
-            break;
-        case "E":
-            // Environment variables update - directly update block metadata
-            if (cmd.data && Object.keys(cmd.data).length > 0) {
-                setTimeout(() => {
-                    fireAndForget(async () => {
-                        await RpcApi.SetMetaCommand(TabRpcClient, {
-                            oref: WOS.makeORef("block", blockId),
-                            meta: { "cmd:env": cmd.data },
-                        }).catch((e) => console.log("error setting cmd:env (OSC 16162 E)", e));
-                    });
-                }, 0);
-
-                // Check for AGENTMUX_AGENT_ID and register/unregister with reactive backend
-                const agentId = cmd.data["AGENTMUX_AGENT_ID"] as string | undefined;
-                const tabId = globalStore.get(atoms.staticTabId);
-                handleAgentIdChange(blockId, agentId, tabId);
-            }
-            break;
-        case "X":
-            // AgentMux configuration - update poller config at runtime
-            fireAndForget(async () => {
-                try {
-                    const url = getWebServerEndpoint() + "/wave/reactive/poller/config";
-                    const response = await fetch(url, {
-                        method: "POST",
-                        headers: { "Content-Type": "application/json" },
-                        body: JSON.stringify({
-                            agentmux_url: cmd.data.agentmux_url || "",
-                            agentmux_token: cmd.data.agentmux_token || "",
-                        }),
-                    });
-                    if (!response.ok) {
-                        const data = await response.json();
-                        console.error("[reactive] failed to configure agentmux:", data.error || response.status);
-                    } else {
-                        const data = await response.json();
-                        console.log("[reactive] agentmux configured:", data.running ? "running" : "stopped");
-                    }
-                } catch (e) {
-                    console.error("[reactive] error configuring agentmux:", e);
-                }
-            });
-            break;
-    }
-
-    if (Object.keys(rtInfo).length > 0) {
-        setTimeout(() => {
-            fireAndForget(async () => {
-                const rtInfoData: CommandSetRTInfoData = {
-                    oref: WOS.makeORef("block", blockId),
-                    data: rtInfo,
-                };
-                await RpcApi.SetRTInfoCommand(TabRpcClient, rtInfoData).catch((e) =>
-                    console.log("error setting RT info (OSC 16162)", e)
-                );
-            });
-        }, 0);
-    }
-
-    return true;
-}
-
+/**
+ * TermWrap — xterm.js wrapper with a strict 3-phase lifecycle.
+ *
+ * Phase 1: CONSTRUCT (sync) — create Terminal, load addons, register OSC handlers.
+ *          NO DOM mount, NO data subscription, NO backend communication.
+ *
+ * Phase 2: INIT (async) — mount to DOM, subscribe to data stream, load initial data,
+ *          flush buffered data, THEN resync controller (which spawns the PTY).
+ *          This ordering eliminates the race condition where PTY output arrives
+ *          before the frontend is ready to receive it.
+ *
+ * Phase 3: RUNNING — handleResize, receive data, periodic cache.
+ */
 export class TermWrap {
     blockId: string;
     ptyOffset: number;
@@ -463,6 +83,8 @@ export class TermWrap {
     pasteActive: boolean = false;
     lastUpdated: number;
 
+    // ── Phase 1: CONSTRUCT (sync) ──────────────────────────────────────
+
     constructor(
         blockId: string,
         connectElem: HTMLDivElement,
@@ -474,8 +96,14 @@ export class TermWrap {
         this.sendDataHandler = waveOptions.sendDataHandler;
         this.ptyOffset = 0;
         this.dataBytesProcessed = 0;
-        this.hasResized = false;
         this.lastUpdated = Date.now();
+        this.connectElem = connectElem;
+        this.mainFileSubject = null;
+        this.heldData = [];
+        this.hasResized = false;
+        this.handleResize_debounced = debounce(50, this.handleResize.bind(this));
+
+        // Create terminal and load addons
         this.terminal = new Terminal(options);
         this.fitAddon = new FitAddon();
         this.fitAddon.noScrollbar = PLATFORM === PlatformMacOS;
@@ -484,8 +112,7 @@ export class TermWrap {
         this.terminal.loadAddon(this.searchAddon);
         this.terminal.loadAddon(this.fitAddon);
         this.terminal.loadAddon(this.serializeAddon);
-        const unicodeAddon = new UnicodeGraphemesAddon();
-        this.terminal.loadAddon(unicodeAddon);
+        this.terminal.loadAddon(new UnicodeGraphemesAddon());
         this.terminal.loadAddon(
             new WebLinksAddon((e, uri) => {
                 e.preventDefault();
@@ -503,37 +130,9 @@ export class TermWrap {
         this.terminal.registerLinkProvider(
             new FilePathLinkProvider(this.terminal, makeFilePathHandler(getCwd))
         );
-        if (WebGLSupported && waveOptions.useWebGl) {
-            try {
-                const webglAddon = new WebglAddon();
-                this.toDispose.push(
-                    webglAddon.onContextLoss(() => {
-                        webglAddon.dispose();
-                        // Fallback to Canvas renderer on WebGL context loss
-                        console.warn("WebGL context lost, falling back to Canvas renderer");
-                        const canvasAddon = new CanvasAddon();
-                        this.toDispose.push(canvasAddon);
-                        this.terminal.loadAddon(canvasAddon);
-                    })
-                );
-                this.terminal.loadAddon(webglAddon);
-                if (!loggedWebGL) {
-                    console.log("loaded webgl renderer!");
-                    loggedWebGL = true;
-                }
-            } catch (e) {
-                // WebGL addon failed to load (common on macOS WKWebView in Tauri)
-                console.warn("WebGL renderer unavailable, using Canvas renderer:", e);
-                const canvasAddon = new CanvasAddon();
-                this.toDispose.push(canvasAddon);
-                this.terminal.loadAddon(canvasAddon);
-                if (!loggedWebGL) {
-                    console.log("loaded canvas renderer (webgl fallback)!");
-                    loggedWebGL = true;
-                }
-            }
-        }
-        // Register OSC 9283 handler
+        this.loadRendererAddon(waveOptions.useWebGl);
+
+        // Register OSC handlers
         this.terminal.parser.registerOscHandler(9283, (data: string) => {
             return handleOscWaveCommand(data, this.blockId, this.loaded);
         });
@@ -543,7 +142,6 @@ export class TermWrap {
         this.terminal.parser.registerOscHandler(16162, (data: string) => {
             return handleOsc16162Command(data, this.blockId, this.loaded, this.terminal);
         });
-        // OSC 0 and 2 - Window title (Claude Code activity summaries)
         this.terminal.parser.registerOscHandler(0, (data: string) => {
             return handleOscTitleCommand(data, this.blockId, this.loaded);
         });
@@ -551,28 +149,20 @@ export class TermWrap {
             return handleOscTitleCommand(data, this.blockId, this.loaded);
         });
         this.terminal.attachCustomKeyEventHandler(waveOptions.keydownHandler);
-        this.connectElem = connectElem;
-        this.mainFileSubject = null;
-        this.heldData = [];
-        this.handleResize_debounced = debounce(50, this.handleResize.bind(this));
-        this.terminal.open(this.connectElem);
-        this.handleResize();
-        let pasteEventHandler = () => {
-            this.pasteActive = true;
-            setTimeout(() => {
-                this.pasteActive = false;
-            }, 30);
-        };
-        pasteEventHandler = pasteEventHandler.bind(this);
-        this.connectElem.addEventListener("paste", pasteEventHandler, true);
-        this.toDispose.push({
-            dispose: () => {
-                this.connectElem.removeEventListener("paste", pasteEventHandler, true);
-            },
-        });
     }
 
-    async initTerminal() {
+    // ── Phase 2: INIT (async) ──────────────────────────────────────────
+
+    /**
+     * Initialize the terminal with correct ordering to prevent race conditions.
+     * Sequence: mount → subscribe → load data → flush held → resync controller.
+     */
+    async init() {
+        // Mount terminal to DOM
+        this.terminal.open(this.connectElem);
+        this.setupPasteHandler();
+
+        // Register input handlers
         const copyOnSelectAtom = getSettingsKeyAtom("term:copyonselect");
         this.toDispose.push(this.terminal.onData(this.handleTermData.bind(this)));
         this.toDispose.push(this.terminal.onKey(this.onKeyHandler.bind(this)));
@@ -592,33 +182,47 @@ export class TermWrap {
         if (this.onSearchResultsDidChange != null) {
             this.toDispose.push(this.searchAddon.onDidChangeResults(this.onSearchResultsDidChange.bind(this)));
         }
+
+        // Subscribe to PTY data stream BEFORE any backend communication.
+        // This ensures we never miss data from the PTY.
         this.mainFileSubject = getFileSubject(this.blockId, TermFileName);
         this.mainFileSubject.subscribe(this.handleNewFileSubjectData.bind(this));
+
+        // Load any existing terminal data (cache + main file)
         try {
             await this.loadInitialTerminalData();
         } finally {
+            // Flush any data that arrived during loading, then open the gate
+            this.flushHeldData();
             this.loaded = true;
         }
+
+        // NOW fit and tell backend to start/resync the shell controller.
+        // At this point we are fully subscribed and ready to receive data.
+        this.fitAddon.fit();
+        this.sendTermSize();
+        await this.resyncController("init");
+        this.hasResized = true;
+
         this.runProcessIdleTimeout();
     }
 
+    // ── Phase 3: RUNNING ───────────────────────────────────────────────
+
     dispose() {
-        // Unregister agent if this block had one registered
         const agentId = registeredAgentsByBlock.get(this.blockId);
         if (agentId) {
             fireAndForget(() => unregisterAgent(agentId));
             registeredAgentsByBlock.delete(this.blockId);
         }
-
-        // Dispose subscriptions and callbacks FIRST - they may reference terminal addons
         this.toDispose.forEach((d) => {
             try {
                 d.dispose();
             } catch (_) {}
         });
-        this.mainFileSubject.release();
-        // Dispose terminal LAST - this also disposes its addons internally
-        // Wrap in try-catch to prevent xterm addon disposal errors from crashing React
+        if (this.mainFileSubject) {
+            this.mainFileSubject.release();
+        }
         try {
             this.terminal.dispose();
         } catch (e) {
@@ -684,7 +288,98 @@ export class TermWrap {
         return prtn;
     }
 
-    async loadInitialTerminalData(): Promise<void> {
+    handleResize() {
+        const oldRows = this.terminal.rows;
+        const oldCols = this.terminal.cols;
+        this.fitAddon.fit();
+        if (oldRows !== this.terminal.rows || oldCols !== this.terminal.cols) {
+            this.sendTermSize();
+        }
+        dlog("resize", `${this.terminal.rows}x${this.terminal.cols}`, `${oldRows}x${oldCols}`);
+    }
+
+    // ── Private helpers ────────────────────────────────────────────────
+
+    private loadRendererAddon(useWebGl: boolean) {
+        if (WebGLSupported && useWebGl) {
+            try {
+                const webglAddon = new WebglAddon();
+                this.toDispose.push(
+                    webglAddon.onContextLoss(() => {
+                        webglAddon.dispose();
+                        console.warn("WebGL context lost, falling back to Canvas renderer");
+                        const canvasAddon = new CanvasAddon();
+                        this.toDispose.push(canvasAddon);
+                        this.terminal.loadAddon(canvasAddon);
+                    })
+                );
+                this.terminal.loadAddon(webglAddon);
+                if (!loggedWebGL) {
+                    console.log("loaded webgl renderer!");
+                    loggedWebGL = true;
+                }
+            } catch (e) {
+                console.warn("WebGL renderer unavailable, using Canvas renderer:", e);
+                const canvasAddon = new CanvasAddon();
+                this.toDispose.push(canvasAddon);
+                this.terminal.loadAddon(canvasAddon);
+                if (!loggedWebGL) {
+                    console.log("loaded canvas renderer (webgl fallback)!");
+                    loggedWebGL = true;
+                }
+            }
+        }
+    }
+
+    private setupPasteHandler() {
+        let pasteEventHandler = () => {
+            this.pasteActive = true;
+            setTimeout(() => {
+                this.pasteActive = false;
+            }, 30);
+        };
+        pasteEventHandler = pasteEventHandler.bind(this);
+        this.connectElem.addEventListener("paste", pasteEventHandler, true);
+        this.toDispose.push({
+            dispose: () => {
+                this.connectElem.removeEventListener("paste", pasteEventHandler, true);
+            },
+        });
+    }
+
+    private flushHeldData() {
+        for (const data of this.heldData) {
+            this.doTerminalWrite(data, null);
+        }
+        this.heldData = [];
+    }
+
+    private sendTermSize() {
+        const termSize: TermSize = { rows: this.terminal.rows, cols: this.terminal.cols };
+        const wsCommand: SetBlockTermSizeWSCommand = {
+            wscommand: "setblocktermsize",
+            blockid: this.blockId,
+            termsize: termSize,
+        };
+        sendWSCommand(wsCommand);
+    }
+
+    async resyncController(reason: string) {
+        dlog("resync controller", this.blockId, reason);
+        const tabId = globalStore.get(atoms.staticTabId);
+        const rtOpts: RuntimeOpts = { termsize: { rows: this.terminal.rows, cols: this.terminal.cols } };
+        try {
+            await RpcApi.ControllerResyncCommand(TabRpcClient, {
+                tabid: tabId,
+                blockid: this.blockId,
+                rtopts: rtOpts,
+            });
+        } catch (e) {
+            console.log(`error controller resync (${reason})`, this.blockId, e);
+        }
+    }
+
+    private async loadInitialTerminalData(): Promise<void> {
         let startTs = Date.now();
         const { data: cacheData, fileInfo: cacheFile } = await fetchWaveFile(this.blockId, TermCacheFileName);
         let ptyOffset = 0;
@@ -717,41 +412,6 @@ export class TermWrap {
         }
     }
 
-    async resyncController(reason: string) {
-        dlog("resync controller", this.blockId, reason);
-        const tabId = globalStore.get(atoms.staticTabId);
-        const rtOpts: RuntimeOpts = { termsize: { rows: this.terminal.rows, cols: this.terminal.cols } };
-        try {
-            await RpcApi.ControllerResyncCommand(TabRpcClient, {
-                tabid: tabId,
-                blockid: this.blockId,
-                rtopts: rtOpts,
-            });
-        } catch (e) {
-            console.log(`error controller resync (${reason})`, this.blockId, e);
-        }
-    }
-
-    handleResize() {
-        const oldRows = this.terminal.rows;
-        const oldCols = this.terminal.cols;
-        this.fitAddon.fit();
-        if (oldRows !== this.terminal.rows || oldCols !== this.terminal.cols) {
-            const termSize: TermSize = { rows: this.terminal.rows, cols: this.terminal.cols };
-            const wsCommand: SetBlockTermSizeWSCommand = {
-                wscommand: "setblocktermsize",
-                blockid: this.blockId,
-                termsize: termSize,
-            };
-            sendWSCommand(wsCommand);
-        }
-        dlog("resize", `${this.terminal.rows}x${this.terminal.cols}`, `${oldRows}x${oldCols}`, this.hasResized);
-        if (!this.hasResized) {
-            this.hasResized = true;
-            this.resyncController("initial resize");
-        }
-    }
-
     processAndCacheData() {
         if (this.dataBytesProcessed < MinDataProcessedForCache) {
             return;
@@ -765,7 +425,7 @@ export class TermWrap {
         this.dataBytesProcessed = 0;
     }
 
-    runProcessIdleTimeout() {
+    private runProcessIdleTimeout() {
         setTimeout(() => {
             if (typeof window.requestIdleCallback === "function") {
                 window.requestIdleCallback(() => {
